# [PERFORMANCE] Limit context switches