--- a/server/protocols/jmap-draft/src/main/java/org/apache/james/jmap/event/ComputeMessageFastViewProjectionListener.java
+++ b/server/protocols/jmap-draft/src/main/java/org/apache/james/jmap/event/ComputeMessageFastViewProjectionListener.java
@@ -44,7 +44,6 @@ import com.google.common.annotations.VisibleForTesting;
 
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import reactor.core.scheduler.Schedulers;
 
 public class ComputeMessageFastViewProjectionListener implements EventListener.ReactiveGroupEventListener {
     public static class ComputeMessageFastViewProjectionListenerGroup extends Group {
@@ -90,9 +89,8 @@ public class ComputeMessageFastViewProjectionListener implements EventListener.R
     private Mono<Void> handleAddedEvent(Added addedEvent, MailboxSession session) {
         return Flux.from(messageIdManager.getMessagesReactive(addedEvent.getMessageIds(), FetchGroup.FULL_CONTENT, session))
             .flatMap(Throwing.function(messageResult -> Mono.fromCallable(
-                () -> Pair.of(messageResult.getMessageId(), computeFastViewPrecomputedProperties(messageResult)))
-                    .subscribeOn(Schedulers.parallel())), DEFAULT_CONCURRENCY)
-            .publishOn(Schedulers.elastic())
+                () -> Pair.of(messageResult.getMessageId(),
+                    computeFastViewPrecomputedProperties(messageResult)))), DEFAULT_CONCURRENCY)
             .flatMap(message -> messageFastViewProjection.store(message.getKey(), message.getValue()), DEFAULT_CONCURRENCY)
             .then();
     }